+++ src/main/java/life/leeray/community/controller/QuestionController.java
@@ -20,7 +20,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id, Model model) {
+    public String question(@PathVariable("id") Long id, Model model) {
         questionService.incView(id);
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("question", question);
